requirements.txt

groq==0.9.0
langchain==0.2.0
langchain-huggingface==0.0.3
langchain-groq==0.1.9 
sqlalchemy==2.0.31
python-dotenv==1.0.1
chromadb==0.5.5
pypdf2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
streamlit==1.37.0